+++ framework/helper/logging.h
@@ -7,19 +7,19 @@
 #include "library/data_structures/queue.h"
 #include "config/game-config.h"
 
-const uint LOG_LEVEL_INFO = 0;
-const uint LOG_LEVEL_WARNING = 1;
-const uint LOG_LEVEL_ERROR = 2;
+const size_t LOG_LEVEL_INFO = 0;
+const size_t LOG_LEVEL_WARNING = 1;
+const size_t LOG_LEVEL_ERROR = 2;
 
-void INTERNAL_T3_Log(uint logLevel, const char *filename, int lineNumber, const char *message, ...);
+void INTERNAL_T3_Log(size_t logLevel, const char *filename, int lineNumber, const char *message, ...);
 
-void INTERNAL_T3_Log(uint logLevel, const char *filename, int lineNumber, T3_LinkedList *linkedList, ...);
+void INTERNAL_T3_Log(size_t logLevel, const char *filename, int lineNumber, T3_LinkedList *linkedList, ...);
 
-void INTERNAL_T3_Log(uint logLevel, const char *filename, int lineNumber, T3_LinkedListDouble *linkedList, ...);
+void INTERNAL_T3_Log(size_t logLevel, const char *filename, int lineNumber, T3_LinkedListDouble *linkedList, ...);
 
-void INTERNAL_T3_Log(uint logLevel, const char *filename, int lineNumber, T3_Stack *stack, ...);
+void INTERNAL_T3_Log(size_t logLevel, const char *filename, int lineNumber, T3_Stack *stack, ...);
 
-void INTERNAL_T3_Log(uint logLevel, const char *filename, int lineNumber, T3_Queue *queue, ...);
+void INTERNAL_T3_Log(size_t logLevel, const char *filename, int lineNumber, T3_Queue *queue, ...);
 
 #define T3_Log(level, message, ...) \
 INTERNAL_T3_Log(level, __FILE__, __LINE__, message, ## __VA_ARGS__)
